Updated algorithm to support promotions (with single product). Tests written passed.

+++ PromotionEngine.Domains/Promotion.cs
@@ -7,6 +7,6 @@ namespace PromotionEngine.Domains
     public class Promotion
     {
         public Dictionary<string, int> ProductAndQuantity { get; set; }
-        public double PromotionPrice { get; set; }
+        public float PromotionPrice { get; set; }
     }
 }
